basis for user authentication
losely based on these resources:

http://nithinbekal.com/posts/phoenix-authentication/
http://meatherly.github.io/2015/05/11/phoenixauthentication/

kills default page_* files.

used the example in web/static/js/socket.js losely for token
authentication. In fact the main difference is that the
Yggdrasil.Authenticate module is used to inject the token into the
conn's assign other than that it's pretty much the same.

the token is availabe in JS via window.userToken, please look at the
example mentioned above for details.

@copenhas delayed socket initialization to avoid errors when ther user
isn't logged in and not on /client.

tests for Yggdrasil.Authenticate should be fairly complete.

tests for Yggdrasil.User exercise a lot of different cases, and for now
Yggdrasil.User is really only setup for inserts, and instead of the
default changeset function create_changeset was used. Going forward this
will allow us to add changesets for update etc.