refactored token auth for channels server only
Found that socket.js has an excellent example of how to implement this
and followed roughly followed it. I modified the Authenticate plug to
inject the token into the conn's assigns map then followed the rest of
the example.

When a user goes to an authenticated page that uses the Authenticate
plug a token will be inserted into window.userToken. For example if the
user clicks refresh a new token will be issued. Probably better ways to
optimize this but for now it works and gets us started.