Client changes for token auth refactor
Simply added a toke property to the params object being passed to
connect. As stated in previous commit the token is stored on
window.userToken.

This actually works as is, since the whole page gets rerendered moving
from login to /client you will see errors in the logs on both client and
server until you log in and once the browser navigates to /client you
will see it say joined "game:lobby" and you can use the client.

I think @copenhas will be working on having this only connect
when you navigate to /client which will be much cleaner.