handle clients joining, welcome them as they connect, and track whose online
Clients can join the player channel on a topic only for that user. The
system keeps track of which users are online via a Player.Registry
and ensures duplicate player processes are not started. Also player
processes are told at spawn time how to send messages to the client by
the channel. When a player processes successfully spawns and registers,
the user reiceves a welcome message. If the client disconnects the
player process will exit gracefully.