Adding emberjs to replace phoenix html in master
Phoenix changes:

Added CORSPlug, check the mix.exs file and plugged it in just before the
router in Yggdrasil.Endpoint, very easy to spot and was suggested by the
projects README.

Ember changes:

uses:

- ember-simple-auth addon which manages the client side session,
and largely followed their example app here[1] as an example of what was
needed to wire it in.

[1] https://github.com/simplabs/ember-simple-auth/tree/master/tests/dummy/app

- ember-bootstrap-less for styling

- phoenix.js which was copied from master and put in app/utils because
  ember cli doesn't transpile vendor

added:

- Login-form a component that handles the login stuff and calls the
  sesion to login
- Registration-form similar to login-form but for registration
- game-client a console like interface to the game
- Index
- yggdrasil.js implements authenticator stuff for ember-cli-auth
- session-manager that handles login/register and retreiving auth token
- game-channel is an alter version of the old socket.js from master
  manages the web socket

Dev scripts:

Updated dev tmux script to include a tab for ember serve.

The two tabs for server stuff now run the commands to fetch deps for
each one before running just to ensure they all is well.