basic command parsing and prototype of execution
Commands are defined as a struct with the related data from the
command text. Each struct must implement the Command protocol.

Command parsing is generated via macros similar to Phoenix routes. The
placeholders from the command pattern text are used to populate the
fields on the struct you assign.

Command execute is shown as passing in a context struct and providing
a module of functions to allow commands to manipulate the struct finally
returning it. This means that command execution is side effect free and
should be fully testable. Another gain is that the  module of functions
gives a means to restrict what commands are allowed to do, as well as
being able to replicate the API in another language for game object
scripting (ex: Lua).